Move DeviceFiller to core and use it in Tango
