+++ src/ophyd_async/core/_hdf_dataset.py
@@ -48,16 +48,14 @@ class HDFFile:
 
         bundler_composer = ComposeStreamResource()
 
-        uri = urlunparse(
-            (
-                "file",
-                self._hostname,
-                str(full_file_name.absolute()),
-                "",
-                "",
-                None,
-            )
-        )
+        uri = urlunparse((
+            "file",
+            self._hostname,
+            str(full_file_name.absolute()),
+            "",
+            "",
+            None,
+        ))
 
         self._bundles: list[ComposeStreamResourceBundle] = [
             bundler_composer(
