+++ src/ophyd_async/tango/__init__.py
@@ -7,14 +7,13 @@ from .signal import (
     AttributeProxy,
     CommandProxy,
     TangoSignalBackend,
-    __tango_signal_auto,
     ensure_proper_executor,
     get_dtype_extended,
     get_python_type,
     get_tango_trl,
     get_trl_descriptor,
     infer_python_type,
-    infer_signal_character,
+    infer_signal_type,
     make_backend,
     tango_signal_r,
     tango_signal_rw,
@@ -32,12 +31,11 @@ __all__ = [
     "get_trl_descriptor",
     "get_tango_trl",
     "infer_python_type",
-    "infer_signal_character",
+    "infer_signal_type",
     "make_backend",
     "AttributeProxy",
     "CommandProxy",
     "ensure_proper_executor",
-    "__tango_signal_auto",
     "tango_signal_r",
     "tango_signal_rw",
     "tango_signal_w",
