+++ src/ophyd_async/tango/signal/_signal.py
@@ -2,21 +2,28 @@
 
 from __future__ import annotations
 
+import logging
 from enum import Enum, IntEnum
 
 import numpy.typing as npt
 
-from ophyd_async.core import DEFAULT_TIMEOUT, SignalR, SignalRW, SignalW, SignalX, T
-from ophyd_async.tango.signal._tango_transport import (
-    TangoSignalBackend,
-    get_python_type,
+from ophyd_async.core import (
+    DEFAULT_TIMEOUT,
+    Signal,
+    SignalDatatypeT,
+    SignalR,
+    SignalRW,
+    SignalW,
+    SignalX,
 )
 from tango import AttrDataFormat, AttrWriteType, CmdArgType, DeviceProxy, DevState
 from tango.asyncio import DeviceProxy as AsyncDeviceProxy
 
+from ._tango_transport import TangoSignalBackend, get_python_type
+
 
 def make_backend(
-    datatype: type[T] | None,
+    datatype: type[SignalDatatypeT] | None,
     read_trl: str = "",
     write_trl: str = "",
     device_proxy: DeviceProxy | None = None,
@@ -25,13 +32,13 @@ def make_backend(
 
 
 def tango_signal_rw(
-    datatype: type[T],
+    datatype: type[SignalDatatypeT],
     read_trl: str,
     write_trl: str = "",
     device_proxy: DeviceProxy | None = None,
     timeout: float = DEFAULT_TIMEOUT,
     name: str = "",
-) -> SignalRW[T]:
+) -> SignalRW[SignalDatatypeT]:
     """Create a `SignalRW` backed by 1 or 2 Tango Attribute/Command
 
     Parameters
@@ -54,12 +61,12 @@ def tango_signal_rw(
 
 
 def tango_signal_r(
-    datatype: type[T],
+    datatype: type[SignalDatatypeT],
     read_trl: str,
     device_proxy: DeviceProxy | None = None,
     timeout: float = DEFAULT_TIMEOUT,
     name: str = "",
-) -> SignalR[T]:
+) -> SignalR[SignalDatatypeT]:
     """Create a `SignalR` backed by 1 Tango Attribute/Command
 
     Parameters
@@ -80,12 +87,12 @@ def tango_signal_r(
 
 
 def tango_signal_w(
-    datatype: type[T],
+    datatype: type[SignalDatatypeT],
     write_trl: str,
     device_proxy: DeviceProxy | None = None,
     timeout: float = DEFAULT_TIMEOUT,
     name: str = "",
-) -> SignalW[T]:
+) -> SignalW[SignalDatatypeT]:
     """Create a `SignalW` backed by 1 Tango Attribute/Command
 
     Parameters
@@ -128,39 +135,10 @@ def tango_signal_x(
     return SignalX(backend, timeout=timeout, name=name)
 
 
-async def __tango_signal_auto(
-    datatype: type[T] | None = None,
-    *,
-    trl: str,
-    device_proxy: DeviceProxy | None,
-    timeout: float = DEFAULT_TIMEOUT,
-    name: str = "",
-) -> SignalW | SignalX | SignalR | SignalRW | None:
-    try:
-        signal_character = await infer_signal_character(trl, device_proxy)
-    except RuntimeError as e:
-        if "Commands with different in and out dtypes" in str(e):
-            return None
-        else:
-            raise e
-
-    if datatype is None:
-        datatype = await infer_python_type(trl, device_proxy)
-
-    backend = make_backend(datatype, trl, trl, device_proxy)
-    if signal_character == "RW":
-        return SignalRW(backend=backend, timeout=timeout, name=name)
-    if signal_character == "R":
-        return SignalR(backend=backend, timeout=timeout, name=name)
-    if signal_character == "W":
-        return SignalW(backend=backend, timeout=timeout, name=name)
-    if signal_character == "X":
-        return SignalX(backend=backend, timeout=timeout, name=name)
-
-
 async def infer_python_type(
     trl: str = "", proxy: DeviceProxy | None = None
 ) -> object | npt.NDArray | type[DevState] | IntEnum:
+    # TODO: work out if this is still needed
     device_trl, tr_name = trl.rsplit("/", 1)
     if proxy is None:
         dev_proxy = await AsyncDeviceProxy(device_trl)
@@ -187,7 +165,9 @@ async def infer_python_type(
     return npt.NDArray[py_type] if isarray else py_type
 
 
-async def infer_signal_character(trl, proxy: DeviceProxy | None = None) -> str:
+async def infer_signal_type(
+    trl, proxy: DeviceProxy | None = None
+) -> type[Signal] | None:
     device_trl, tr_name = trl.rsplit("/", 1)
     if proxy is None:
         dev_proxy = await AsyncDeviceProxy(device_trl)
@@ -204,20 +184,19 @@ async def infer_signal_character(trl, proxy: DeviceProxy | None = None) -> str:
     if tr_name in dev_proxy.get_attribute_list():
         config = await dev_proxy.get_attribute_config(tr_name)
         if config.writable in [AttrWriteType.READ_WRITE, AttrWriteType.READ_WITH_WRITE]:
-            return "RW"
+            return SignalRW
         elif config.writable == AttrWriteType.READ:
-            return "R"
+            return SignalR
         else:
-            return "W"
+            return SignalW
 
     if tr_name in dev_proxy.get_command_list():
         config = await dev_proxy.get_command_config(tr_name)
         if config.in_type == CmdArgType.DevVoid:
-            return "X"
+            return SignalX
         elif config.in_type != config.out_type:
-            raise RuntimeError(
-                "Commands with different in and out dtypes are not" " supported"
-            )
+            logging.debug("Commands with different in and out dtypes are not supported")
+            return None
         else:
-            return "RW"
+            return SignalRW
     raise RuntimeError(f"Unable to infer signal character for {trl}")
